use long

+++ src/main/java/Main.java
@@ -45,7 +45,7 @@ public class Main {
       ResponseMessage rcvivedMessage = new ResponseMessage( null, null) ;
       java.util.HashMap<String, byte[]> headerMap = new java.util.HashMap<>();
       ByteBuffer buffer = ByteBuffer.allocate(8);
-      buffer.putInt(7);
+      buffer.putLong(7);
       headerMap.put("correlation_id", buffer.array());
       rcvivedMessage.setHeader(headerMap);
       rcvivedMessage.setMessage_size(5);
